#17 productList 위치 수정
