+++ KnockKnock/src/main/java/com/knockknock/mypage/MypageProductService.java
@@ -7,11 +7,7 @@ import com.knockknock.user.UserVO;
 import com.knockknock.util.ProductVO;
 
 public interface MypageProductService {
-	// 상품 목록으로 이동
-	List<ProductVO> ProductList();
-	
-	// 상품정보 가져오기(임시
-	ProductVO productDetail(int pIdx);
+
 	
 	// 장바구니에 상품 담기 전 동일한 상품번호 있는지 확인
 	int checkCart(int pIdx);
