+++ KnockKnock/src/main/java/com/knockknock/mypage/mypageproductimpl/MypageProductDAO.java
@@ -15,18 +15,7 @@ public class MypageProductDAO {
 	
 	@Autowired
 	private SqlSessionTemplate mybatis;
-	
-	// 상품 목록으로 이동(임시)
-	public List<ProductVO> ProductList() {
 
-		return mybatis.selectList("UserVO.productList");
-	}
-
-	// 상품 정보 가져오기(임시)
-	public ProductVO productDetail(int pIdx) {
-
-		return mybatis.selectOne("UserVO.productDetail", pIdx);
-	}
 	// 장바구니에 상품 담기 전 동일한 상품 있는지 확인
 	public int checkCart(int pIdx) {
 		System.out.println("pIdx : " + pIdx);
