+++ KnockKnock/src/main/java/com/knockknock/mypage/mypageproductimpl/MypageProductServiceImpl.java
@@ -16,18 +16,7 @@ public class MypageProductServiceImpl implements MypageProductService {
 	@Autowired
 	private MypageProductDAO mypageProductDAO;
 	
-	// 상품 목록으로 이동(임시)
-	@Override
-	public List<ProductVO> ProductList() {
-		return mypageProductDAO.ProductList();
-	}
 
-	// 상품정보 가져오기 (임시)
-	@Override
-	public ProductVO productDetail(int pIdx) {
-		
-		return mypageProductDAO.productDetail(pIdx);
-	}
 	// 장바구니에 상품 담기 전 동일한 상품번호 있는지 확인
 	@Override
 	public int checkCart(int pIdx) {
