+++ KnockKnock/src/main/java/com/knockknock/product/ProductService.java
@@ -1,10 +1,13 @@
 package com.knockknock.product;
 
-import java.util.List;
+import java.util.List; 
 
-import com.knockknock.util.ProductVO;
+import com.knockknock.product.ProductVO;
 
 public interface ProductService {
+	// 상품목록
 	List<ProductVO> ProductList();
 
+	// 상품정보 가져오기(임시
+	ProductVO productDetail(int pIdx);
 }
