+++ KnockKnock/src/main/java/com/knockknock/product/impl/ProductServiceImpl.java
@@ -1,12 +1,12 @@
 package com.knockknock.product.impl;
 
-import java.util.List;
+import java.util.List; 
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.knockknock.product.ProductService;
-import com.knockknock.util.ProductVO;
+import com.knockknock.product.ProductVO;
 
 @Service("productService")
 public class ProductServiceImpl implements ProductService{
@@ -16,5 +16,12 @@ public class ProductServiceImpl implements ProductService{
 	@Override
 	public List<ProductVO> ProductList() {
 		return productDAO.ProductList();
+	}
+	
+	// 상품정보 가져오기 (임시)
+	@Override
+	public ProductVO productDetail(int pIdx) {
+		
+		return productDAO.productDetail(pIdx);
 	}
 }
